+++ examples/apps/branching/src/healthBot.ts
@@ -5,6 +5,12 @@
 
 import type { GroceryListJSON } from "./groceryList/index.js";
 
+/**
+ * Simulating the case that we're making a network request to get suggestions for edits to the content,
+ * as this is a likely scenario.  This is to force us to understand the implications of extracting the data
+ * from Fluid, and later trying to merge suggestions back in (including scenarios where the Fluid content
+ * has changed in the meantime).
+ */
 export const NETWORK_askHealthBotForSuggestions = async (
 	groceryListJSONString: string,
 ): Promise<string> => {
